Move registries to API since it's necessary they live there for access in 3rd party plugins
