+++ api/src/main/java/org/reprogle/honeypot/Registry.java
@@ -1,0 +1,38 @@
+/*
+ * Honeypot is a plugin written for Paper which assists with griefing auto-moderation
+ *
+ * Copyright TerrorByte & Honeypot Contributors (c) 2022 - 2024
+ *
+ * This program is free software: You can redistribute it and/or modify it under the terms of the Mozilla Public License 2.0
+ * as published by the Mozilla under the Mozilla Foundation.
+ *
+ * This program is distributed in the hope that it will be useful, but provided on an "as is" basis,
+ * without warranty of any kind, either expressed, implied, or statutory, including, without limitation,
+ * warranties that the Covered Software is free of defects, merchantable, fit for a particular purpose or non-infringing.
+ * See the MPL 2.0 license for more details.
+ *
+ * For a full copy of the license in its entirety, please visit <https://www.mozilla.org/en-US/MPL/2.0/>
+ */
+
+package org.reprogle.honeypot;
+
+public class Registry {
+    private static BehaviorRegistry behaviorRegistry = new BehaviorRegistry();
+    private static StorageManagerRegistry storageManagerRegistry = new StorageManagerRegistry();
+
+    public static BehaviorRegistry getBehaviorRegistry() {
+        return behaviorRegistry;
+    }
+
+    public static void setBehaviorRegistry(BehaviorRegistry behaviorRegistry) {
+        Registry.behaviorRegistry = behaviorRegistry;
+    }
+
+    public static StorageManagerRegistry getStorageManagerRegistry() {
+        return storageManagerRegistry;
+    }
+
+    public static void setStorageManagerRegistry(StorageManagerRegistry storageManagerRegistry) {
+        Registry.storageManagerRegistry = storageManagerRegistry;
+    }
+}
