+++ gradle/libs.versions.toml
@@ -15,7 +15,7 @@ lands = "6.44.6"
 vault = "1.7.1"
 placeholder-api = "2.11.5"
 boosted-yaml = "1.3.5"
-bstats = "3.0.2"
+bstats = "3.1.0"
 
 guice = '7.0.0'
 okhttp = '4.12.0'
